+++ mkosi/config.py
@@ -773,6 +773,7 @@ class MkosiConfig:
     tools_tree_release: Optional[str]
     tools_tree_packages: list[str]
     runtime_trees: list[tuple[Path, Optional[Path]]]
+    runtime_size: Optional[int]
 
     # QEMU-specific options
     qemu_gui: bool
@@ -1759,6 +1760,13 @@ SETTINGS = (
         parse=config_make_list_parser(delimiter=",", parse=make_source_target_paths_parser()),
         help="Additional mounts to add when booting the image",
     ),
+    MkosiConfigSetting(
+        dest="runtime_size",
+        metavar="SIZE",
+        section="Host",
+        parse=config_parse_bytes,
+        help="Grow disk images to the specified size before booting them",
+    ),
 )
 
 MATCHES = (
@@ -2429,6 +2437,21 @@ def line_join_source_target_list(array: Sequence[tuple[Path, Optional[Path]]]) -
     return "\n                                ".join(items)
 
 
+def format_bytes(num_bytes: int) -> str:
+    if num_bytes >= 1024**3:
+        return f"{num_bytes/1024**3 :0.1f}G"
+    if num_bytes >= 1024**2:
+        return f"{num_bytes/1024**2 :0.1f}M"
+    if num_bytes >= 1024:
+        return f"{num_bytes/1024 :0.1f}K"
+
+    return f"{num_bytes}B"
+
+
+def format_bytes_or_none(num_bytes: Optional[int]) -> str:
+    return format_bytes(num_bytes) if num_bytes is not None else "none"
+
+
 def summary(args: MkosiArgs, config: MkosiConfig) -> str:
     def bold(s: Any) -> str:
         return f"{Style.bold}{s}{Style.reset}"
@@ -2560,6 +2583,7 @@ Clean Package Manager Metadata: {yes_no_auto(config.clean_package_metadata)}
             Tools Tree Release: {none_to_none(config.tools_tree_release)}
            Tools Tree Packages: {line_join_list(config.tools_tree_packages)}
                  Runtime Trees: {line_join_source_target_list(config.runtime_trees)}
+                  Runtime Size: {format_bytes_or_none(config.runtime_size)}
 
                       QEMU GUI: {yes_no(config.qemu_gui)}
                 QEMU CPU Cores: {config.qemu_smp}
