+++ mkosi/util.py
@@ -160,17 +160,6 @@ def qemu_check_vsock_support(log: bool) -> bool:
     return True
 
 
-def format_bytes(num_bytes: int) -> str:
-    if num_bytes >= 1024**3:
-        return f"{num_bytes/1024**3 :0.1f}G"
-    if num_bytes >= 1024**2:
-        return f"{num_bytes/1024**2 :0.1f}M"
-    if num_bytes >= 1024:
-        return f"{num_bytes/1024 :0.1f}K"
-
-    return f"{num_bytes}B"
-
-
 def make_executable(path: Path) -> None:
     st = path.stat()
     os.chmod(path, st.st_mode | stat.S_IEXEC)
